+++ substrate/frame/revive/fixtures/contracts/CallerWithConstructor.sol
@@ -8,13 +8,13 @@ contract CallerWithConstructor {
       callee = new CallerWithConstructorCallee();
     }
 
-    function callBar() public view returns (uint) {
+    function callBar() public view returns (uint64) {
       return callee.bar();
     }
 }
 
 contract CallerWithConstructorCallee {
-    function bar() public pure returns (uint) {
+    function bar() public pure returns (uint64) {
       return 42;
     }
 }
